# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Zenythium Fitness API",
        "description": "API для управления тренировками, упражнениями и планами тренировок",
        "contact": {
            "email": "support@zenythium.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Основной сервер API"
        }
    ],
    "paths": {
...
        "/api/v1/user/statistics": {
            "get": {
                "tags": [
                    "Statistics"
                ],
                "summary": "Получение статистики пользователя",
                "description": "Возвращает комплексную статистику тренировок и прогресса пользователя. Включает общее количество тренировок, время тренировок, объемы, изменения веса и частоту тренировок",
                "operationId": "14b070c87e31018911344cd6ea683155",
                "responses": {
                    "200": {
                        "description": "Статистика пользователя успешно получена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "total_workouts": {
                                                    "type": "integer",
                                                    "example": 45
                                                },
                                                "completed_workouts": {
                                                    "type": "integer",
                                                    "example": 42
                                                },
                                                "total_training_time": {
                                                    "type": "integer",
                                                    "example": 3150
                                                },
                                                "total_volume": {
                                                    "type": "integer",
                                                    "example": 125000
                                                },
                                                "current_weight": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 75.5
                                                },
                                                "active_cycles_count": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "weight_change_30_days": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 2.5
                                                },
                                                "training_frequency_4_weeks": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 3.2
                                                },
                                                "training_streak_days": {
                                                    "type": "integer",
                                                    "example": 7
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Статистика пользователя успешно получена"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Пользователь не аутентифицирован"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}